refactor: map 모듈 내 viewmodel, adapter 폴더 생성& 파일 이동
