+++ presentation/user/map/src/main/java/com/binbean/map/adapter/ViewPagerAdapter.kt
@@ -1,9 +1,10 @@
-package com.binbean.map
+package com.binbean.map.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.arieum.review.CafeReviewFragment
 import com.binbean.domain.cafe.Cafe
+import com.binbean.map.CafeDetailedInfoFragment
 
 class ViewPagerAdapter(
     fragment: Fragment,
