+++ presentation/user/map/src/main/java/com/binbean/map/viewmodel/MapViewModel.kt
@@ -1,4 +1,4 @@
-package com.binbean.map
+package com.binbean.map.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.LiveData
@@ -6,7 +6,6 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.binbean.domain.cafe.Cafe
-import com.binbean.domain.cafe.CafeInfoImgItem
 import com.binbean.domain.cafe.SearchCafesInBoundsUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
